--- sedaro/src/sedaro/models/project/ProjectManager.py
@@ -1,10 +1,0 @@
-from typing import ClassVar
-
-from ..BaseModelManager import BaseModelManager
-from .Project import Project
-
-
-class ProjectManager(BaseModelManager[Project]):
-
-    _BASE_PATH: 'ClassVar[str]' = '/projects'
-    _MODEL = Project
