+++ sedaro/src/sedaro/models/project/main.py
@@ -1,6 +1,7 @@
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, ClassVar
 
 from ..BaseModel import BaseModel
+from ..BaseModelManager import BaseModelManager
 
 if TYPE_CHECKING:
     from ..workspace.Workspace import Workspace
@@ -17,3 +18,9 @@ class Project(BaseModel):
         '''Get the workspace of the project.'''
         from ..workspace.Workspace import Workspace
         return self._get_rel('workspace', Workspace)
+
+
+class ProjectManager(BaseModelManager[Project]):
+
+    _BASE_PATH: 'ClassVar[str]' = '/projects'
+    _MODEL = Project
